@@ -6,7 +6,7 @@ authors = ["Suphawit <suphawit.won@gmail.com>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
 flask = "^2.2.2"
 sqlalchemy = "^1.4.46"
 pymysql = "^1.0.2"
