@@ -5,7 +5,7 @@ from flask.testing import FlaskClient
 from werkzeug.test import TestResponse
 
 from app.database.queryUtils import add_obj_to_db
-from app.database.schema import VendingMachine
+from app.database.schema import VendingMachine, Timeline
 from app.tests.conftest import clear_db, get_test_entry, Tester
 
 
@@ -37,6 +37,7 @@ def tester(client: FlaskClient) -> MachineTester:
 def test_view_vending_machine(tester: MachineTester) -> None:
     response = tester.get_all_machines()
     assert Tester.expect(response, 200) or Tester.expect(response, 204)
+    clear_db(Timeline)
     clear_db(VendingMachine)
 
 
@@ -46,6 +47,7 @@ def test_adding_vending_machine(tester: MachineTester) -> None:
     response = tester.add_machines(test_data)
     assert response.status_code == 200 or response.status_code == 204
     assert len(response.json) != 0
+    clear_db(Timeline)
     clear_db(VendingMachine)
 
 
@@ -65,6 +67,7 @@ def test_edit_vending_machine(tester: MachineTester) -> None:
     assert edited_test_entry.id == test_data["id"]
     assert edited_test_entry.machine_name == test_data["machine_name"]
     assert edited_test_entry.machine_location == test_data["machine_location"]
+    clear_db(Timeline)
     clear_db(VendingMachine)
 
 
@@ -75,6 +78,7 @@ def test_delete_vending_machine(tester: MachineTester) -> None:
     response = tester.delete_machines({"id": test_entry.id})
     assert response.status_code == 200 or response.status_code == 204
     assert response.json is None
+    clear_db(Timeline)
     clear_db(VendingMachine)
 
 
@@ -82,6 +86,7 @@ def test_failed_add_vending_machine(tester: MachineTester) -> None:
     clear_db(VendingMachine)
     response = tester.add_machines({})
     assert response.status_code == 400
+    clear_db(Timeline)
     clear_db(VendingMachine)
 
 
@@ -89,4 +94,5 @@ def test_failed_delete_vending_machine(tester: MachineTester) -> None:
     clear_db(VendingMachine)
     response = tester.delete_machines({})
     assert response.status_code == 400
+    clear_db(Timeline)
     clear_db(VendingMachine)
